\name{geostatistics-package}
\alias{geostatistics-package}
\alias{geostatistics}
\docType{package}
\title{
\packageTitle{geostatistics}
}
\description{
\packageDescription{geostatistics}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{geostatistics}
\packageIndices{geostatistics}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{geostatistics}

Maintainer: \packageMaintainer{geostatistics}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
